% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Initialize Connection}
\usage{
connect_to_db(conf_file = "~/.db_conf.yml")
}
\arguments{
\item{conf_file}{(character) Configuration file to read/write.
(Default: "~/.db_conf.yml")}
}
\value{
(DBIConnection) A DBIConnection for success; FALSE for failure.
}
\description{
Initialize a connection to the database and return a DBIConnection.
}
\section{Details}{

A configuration file will be read and used to connect to the database.
}

\examples{
\dontrun{
channel <- connect_to_db()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
