% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_add_col}
\alias{db_add_col}
\title{Add a Column (Field).}
\usage{
db_add_col(tablename, fieldname, fieldtype, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name in a database.}

\item{fieldname}{(character) A field name to add to a database table.}

\item{fieldtype}{(character) A field type to set for a field added to a table.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) The number of rows affected by the field addition.
}
\description{
Add a column (field) to a database table.
}
\section{Details}{

An ALTER TABLE ... ADD COLUMN command will add a field to a table in a
database.
}

\examples{
\dontrun{
db_add_col("iris", "Species", "varchar(32)")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
