% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_append_table}
\alias{db_append_table}
\title{Append a Dataframe to a Table}
\usage{
db_append_table(df, tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{df}{(dataframe) A dataframe to append to a database table.}

\item{tablename}{(character) A table name to receive additional records.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) Number of affected (appended) rows.
}
\description{
Send a dataframe to the database to append to a table.
}
\section{Details}{

A dataframe will be sent to the database to be appended to an existing table.
}

\examples{
\dontrun{
db_append_table(datasets::iris, "iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
