% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_connect}
\alias{db_connect}
\title{Initialize Connection}
\usage{
db_connect(conf_file = "~/.db_conf.yml")
}
\arguments{
\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(DBIConnection) A DBIConnection for success; FALSE for failure.
}
\description{
Initialize a connection to the database and return a DBIConnection.
}
\section{Details}{

A configuration file will be read and used to connect to the database.
}

\examples{
\dontrun{
channel <- db_connect()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
