% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_dim}
\alias{db_dim}
\title{Show dimensions of a table.}
\usage{
db_dim(tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for number of columns.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) A vector containing the count of rows and columns in a table.
}
\description{
Run database queries that returns the number of rows and columns in a table.
}
\section{Details}{

SQL queries will return a vector containing the number of rows and columns.
}

\examples{
\dontrun{
db_dim("iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
