% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_drop_col}
\alias{db_drop_col}
\title{Drop (Delete) a Column (Field).}
\usage{
db_drop_col(tablename, fieldname, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name in a database.}

\item{fieldname}{(character) A field name to remove from a database table.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) The number of rows affected by the field addition.
}
\description{
Remove a column (field) to a database table.
}
\section{Details}{

An ALTER TABLE ... DROP COLUMN command will remove a field from a table in
a database.
}

\examples{
\dontrun{
db_drop_col("iris", "Species")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
