% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_fetch_query}
\alias{db_fetch_query}
\title{Fetch Results from a Query}
\usage{
db_fetch_query(query, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{query}{(character) A SQL statement as a text string.}

\item{conf_file}{(character) Configuration file to read/write.
(Default: "~/.db_conf.yml")}
}
\value{
(dataframe) The query result returned as a dataframe.
}
\description{
Run a database query that returns a dataframe.
}
\section{Details}{

A SQL statement will be run and a dataframe of results will be returned.
}

\examples{
\dontrun{
db_fetch_query("SELECT * FROM my.tablename LIMIT 10;")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
