% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_get_type}
\alias{db_get_type}
\title{Get the Type of a Field.}
\usage{
db_get_type(tablename, fieldname, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for structure.}

\item{fieldname}{(character) A field name to query for Field Type.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(character) The Field Type.
}
\description{
Run a database query on a table that shows the type of a field (column).
}
\section{Details}{

An ALTER TABLE command will return the type of a field in a table in a
database.
}

\examples{
\dontrun{
db_get_type("iris", "Species")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
