% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_len}
\alias{db_len}
\title{Show Length and Size of Tables.}
\usage{
db_len(conf_file = "~/.db_conf.yml")
}
\arguments{
\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(dataframe) The sizes of the tables in a database.
}
\description{
Run a database query that lists the length and size of database tables.
}
\section{Details}{

A SQL query will return the length and sizes of the tables in the database.
}

\examples{
\dontrun{
db_len()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
