% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_ls}
\alias{db_ls}
\title{List tables in a database.}
\usage{
db_ls(conf_file = "~/.db_conf.yml")
}
\arguments{
\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(character) The names of the tables in a database.
}
\description{
Run a database query that lists the tables in a database.
}
\section{Details}{

A SHOW TABLES query will return the names of the tables in the database.
}

\examples{
\dontrun{
db_ls()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
