% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_ncol}
\alias{db_ncol}
\title{Count number of columns of a table.}
\usage{
db_ncol(tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for number of columns.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) The number of columns in a table.
}
\description{
Run a database query that returns the number of columns in a table.
}
\section{Details}{

A SQL query will return a dataframe containing the number of columns.
}

\examples{
\dontrun{
db_ncol("iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
