% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_nrow}
\alias{db_nrow}
\title{Count number of rows of a table.}
\usage{
db_nrow(tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for number of rows.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) The number of rows in a table.
}
\description{
Run a database query that returns the number of rows in a table.
}
\section{Details}{

A SELECT COUNT query will return the number of rows in a table.
}

\examples{
\dontrun{
db_nrow("iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
