% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_read_conf}
\alias{db_read_conf}
\title{Read Configuration}
\usage{
db_read_conf(conf_file = "~/.db_conf.yml")
}
\arguments{
\item{conf_file}{(character) Configuration file to read/write.
(Default: "~/.db_conf.yml")}
}
\description{
Read database configuration file.
}
\section{Details}{

A configuration file will be read if found, otherwise one will be created.
}

\examples{
\dontrun{
db_read_conf()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
