% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_remove_table}
\alias{db_remove_table}
\title{Remove a Table}
\usage{
db_remove_table(tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to remove from the database.}

\item{conf_file}{(character) Configuration file to read/write.
(Default: "~/.db_conf.yml")}
}
\value{
(boolean) Success: TRUE; failure: FALSE.
}
\description{
Remove a table from a database.
}
\section{Details}{

A table will be removed from the database.
}

\examples{
\dontrun{
db_remove_table("iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
