% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_set_type}
\alias{db_set_type}
\title{Set the Type of a Field.}
\usage{
db_set_type(tablename, fieldname, fieldtype, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name in a database.}

\item{fieldname}{(character) A field name in a database table.}

\item{fieldtype}{(character) A field type to set for a field in a table.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(integer) The number of rows affected by the field type change.
}
\description{
Set the type of a field (column) in a database table.
}
\section{Details}{

An ALTER TABLE command will change the type of a field in a table in a
database.
}

\examples{
\dontrun{
db_set_type("iris", "Species", "varchar(32)")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
