% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_str}
\alias{db_str}
\title{Show structure of a table.}
\usage{
db_str(tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for structure.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(dataframe) The query result returned as a dataframe.
}
\description{
Run a database query on a table that shows the columns and their properties.
}
\section{Details}{

A SHOW COLUMNS query will return a dataframe of columns and their properties.
}

\examples{
\dontrun{
db_str("iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
