% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_str_all}
\alias{db_str_all}
\title{Show structure of all tables.}
\usage{
db_str_all(conf_file = "~/.db_conf.yml")
}
\arguments{
\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(dataframe) The combined query results returned as a dataframe.
}
\description{
Show the columns and their properties for all of the tables in a database.
}
\section{Details}{

Results from SHOW COLUMNS queries for all tables in a database will be
combined into a single dataframe.
}

\examples{
\dontrun{
db_str_all()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
