% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_write_conf}
\alias{db_write_conf}
\title{Write Configuration}
\usage{
db_write_conf(
  conf_file = "~/.db_conf.yml",
  username = "",
  host = "",
  dbname = "",
  sslmode = "",
  sslca = "",
  sslkey = "",
  sslcert = ""
)
}
\arguments{
\item{conf_file}{(character) Configuration file to write.
(Default: "~/.db_conf.yml")}

\item{username}{(character) Username.  (Default: "")}

\item{host}{(character) Database erver hostname.
(Default: "")}

\item{dbname}{(character) Database name. (Default: "")}

\item{sslmode}{(character) SSL mode. (Default: "")}

\item{sslca}{(character) CCL CA path. (Default: "")}

\item{sslkey}{(character) SSL key path. (Default: "")}

\item{sslcert}{(character) SSL certificate path. (Default: "")}
}
\description{
Write database configuration settings to a file.
}
\section{Details}{

A configuration file will be created with the values provided. For details
about the arguments, see the documentation for RMariaDB::MariaDB().
}

\examples{
\dontrun{
db_write_conf(conf_file = "~/.db_conf.yml",
             username = "my_username",
             host = "db.server.example.com",
             dbname = "my_dbname",
             sslmode = "REQUIRED",
             sslca = "/etc/db-ssl/ca-cert.pem",
             sslkey = "/etc/db-ssl/client-key-pkcs1.pem",
             sslcert = "/etc/db-ssl/client-cert.pem")
file.edit("~/.db_conf.yml")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
