% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_colnames}
\alias{db_colnames}
\title{Show Column Names for a Table}
\usage{
db_colnames(tablename, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for column (field) names.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(character) A vector of column (field) names.
}
\description{
Show a list of column (field) names for a table in a database.
}
\section{Details}{

The database will be queried for column (field) names for a table. These
names will be returned as a character vector.
}

\examples{
\dontrun{
db_colnames()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
