% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_fetch_table}
\alias{db_fetch_table}
\title{Fetch a Table}
\usage{
db_fetch_table(tablename, n = -1, conf_file = "~/.db_conf.yml")
}
\arguments{
\item{tablename}{(character) A table name to query for all records.}

\item{n}{(integer) The number of rows to return. (Default -1 means all rows.)}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}
}
\value{
(dataframe) The query result returned as a dataframe.
}
\description{
Retrieve a database table as a dataframe.
}
\section{Details}{

A SQL query will return a table as a dataframe. Use 'n' to limit the number
of rows returned.
}

\examples{
\dontrun{
db_fetch_table("iris")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
