% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_read_conf}
\alias{db_read_conf}
\title{Read Configuration}
\usage{
db_read_conf(
  conf_file = "~/.db_conf.yml",
  username = "",
  host = "",
  dbname = "",
  sslmode = "",
  sslca = "",
  sslkey = "",
  sslcert = "",
  enable_cleartext_plugin = TRUE
)
}
\arguments{
\item{conf_file}{(character) Configuration file to read/write.
(Default: "~/.db_conf.yml")}

\item{username}{(character) Username. See: RMariaDB::MariaDB. (Default: "")}

\item{host}{(character) Database erver hostname. See: RMariaDB::MariaDB.
(Default: "")}

\item{dbname}{(character) Database name. See: RMariaDB::MariaDB. (Default: "")}

\item{sslmode}{(character) SSL mode. See: RMariaDB::MariaDB. (Default: "")}

\item{sslca}{(character) CCL CA path. See: RMariaDB::MariaDB. (Default: "")}

\item{sslkey}{(character) SSL key path. See: RMariaDB::MariaDB. (Default: "")}

\item{sslcert}{(character) SSL certificate path. See: RMariaDB::MariaDB.
(Default: "")}

\item{enable_cleartext_plugin}{(boolean) Set the environment variable.
LIBMYSQL_ENABLE_CLEARTEXT_PLUGIN=1. (Default: TRUE)#' @return (boolean) TRUE for success; FALSE for failure.}
}
\description{
Read database configuration file.
}
\section{Details}{

A configuration file will be read if found, otherwise one will be created.
}

\examples{
\dontrun{
# First, run once to create the file with the values provided:
db_read_conf(conf_file = "~/.db_conf.yml",
             username = "my_username",
             host = "db.server.example.com",
             dbname = "my_dbname",
             sslmode = "REQUIRED",
             sslca = "/etc/db-ssl/ca-cert.pem",
             sslkey = "/etc/db-ssl/client-key-pkcs1.pem",
             sslcert = "/etc/db-ssl/client-cert.pem",
             enable_cleartext_plugin = TRUE)
# You will see warnings about the file not existing and/or needs editing.

# Subsequently, read the file once per session:
db_conf <- db_read_conf()
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
