% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDBHelper.R
\name{db_run_query}
\alias{db_run_query}
\title{Run a Query}
\usage{
db_run_query(query, conf_file = "~/.db_conf.yml", ...)
}
\arguments{
\item{query}{(character) A SQL statement as a text string.}

\item{conf_file}{(character) A file containing database connection parameters.
(Default: "~/.db_conf.yml")}

\item{...}{Additional arguments passed to RMariaDB::dbExecute().}
}
\value{
(integer) The number of affected rows.
}
\description{
Run a database query that returns the number of affected rows.
}
\section{Details}{

A SQL statement will be run and the number of affected rows will be returned.
}

\examples{
\dontrun{
db_run_query("DELETE FROM my.tablename WHERE id = 1;")
}
}
\keyword{MariaDB,}
\keyword{database,}
\keyword{sql,}
\keyword{utility}
